Update to how minions are stored with the player and bug fixes.
Minions are now stored as a MonsterGroup instead of an ArrayList.
Minions render their image but doesn't correctly render their healthbar.
Minions are separate form the monster that are in a room to avoid
damaging them with AOE cards or relics and allowing the battle to end
with minions still alive.