Fix ApplyPowerActionPatch and Monster Attacking minion when no minions
ApplyPowerActionPatch wasn't working as intended because you can't
make changes to the parameters unless you use @ByRef but it turned out
there was another way to do it so it's been changed.

Monsters will now damage the player if it was intending to attack a
minion but all minions have died.