Fix when damaging minions and Summoning more minions than player can.
When a base character has a minion and it is attacked the damage is now
directed towards the minion correctly and the player isn't also damaged.

Cards that summon a minion can now be set to unplayable if the player
has no more slots for summoning a minion. This requires that card to
extend CustomSummonCard.

The max amount of minions a player can have can now be increased easily.