Add new way to do minion moves
MinionMoves are now used to hold the information about a move for minions.
This includes a clickable images rendered by the minion that will
set the minions move to that.
Descriptions of the moves are shown when hovered over.

Kept in the old way to use minion moves however it is now marked as
deprecated.